Add TARGET_MODE_CAN_TRANSFER_BITS
The following adds a target hook to specify whether regs of MODE can be
used to transfer bits.  The hook is supposed to be used for value-numbering
to decide whether a value loaded in such mode can be punned to another
mode instead of re-loading the value in the other mode and for SRA to
decide whether MODE is suitable as container holding a value to be
used in different modes.

	* target.def (mode_can_transfer_bits): New target hook.
	* target.h (mode_can_transfer_bits): New function wrapping the
	hook and providing default behavior.
	* doc/tm.texi.in: Update.
	* doc/tm.texi: Re-generate.