match: Fix types matching for `(?:) !=/== (?:)` [PR116134]
The problem here is that in generic types of comparisons don't need
to be boolean types (or vector boolean types). And fixes that by making
sure the types of the conditions match before doing the optimization.

Bootstrapped and tested on x86_64-linux-gnu with no regressions.

	PR middle-end/116134

gcc/ChangeLog:

	* match.pd (`(a ? x : y) eq/ne (b ? x : y)`): Check that
	a and b types match.
	(`(a ? x : y) eq/ne (b ? y : x)`): Likewise.

gcc/testsuite/ChangeLog:

	* gcc.dg/torture/pr116134-1.c: New test.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>