c++, coroutines: Remove unused suspend point state [NFC].
We maintain state on the progress of await analysis in an object that
is passed to the various tree walks used.  Some of the state had become
stale (i.e. unused members).  Remove those and provide a CTOR so that
updates are localised.

Remove the file scope hash_map used to collect the final state for the
actor function and make that part of the suspend point state.

gcc/cp/ChangeLog:

	* coroutines.cc (struct susp_frame_data): Remove unused members,
	provide a CTOR.
	(morph_fn_to_coro): Use susp_frame_data CTOR, and make the suspend
	state hash map local to the morph function.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>