ada: Emit a warning on inheritly limited types
Record types that do not have a limited keyword but have a
member with a limited type are also considered to be limited types.
This can be confusing to understand for newer Ada users. It is
better to emit a warning in this scenario and suggest that the
type should be marked with a limited keyword. This diagnostic will
be acticated when the -gnatw_l switch is used.

gcc/ada/

	* sem_ch3.adb: Add method Check_Inherited_Limted_Record for
	emitting the warning for an inherited limited type.
	* warnsw.adb: Add processing for the -gnatw_l switch that
	triggeres the inheritly limited type warning.
	* warnsw.ads: same as above.
	* doc/gnat_ugn/building_executable_programs_with_gnat.rst: Add
	entry for -gnatw_l switch.
	* gnat_ugn.texi: Regenerate.