c++: Fix up error recovery of invalid structured bindings used in conditions [PR116113]
The following testcase ICEs, because for structured binding error recovery
DECL_DECOMP_BASE is kept NULL and the newly added code to pick up saved
value from the base assumes that on structured binding bases the
TARGET_EXPR will be always there (that is the case if there are no errors).

The following patch fixes it by testing DECL_DECOMP_BASE before
dereferencing it, another option would be not to do that if
error_operand_p (cond).

2024-08-01  Jakub Jelinek  <jakub@redhat.com>

	PR c++/116113
	* semantics.cc (maybe_convert_cond): Check DECL_DECOMP_BASE
	is non-NULL before dereferencing it.
	(finish_switch_cond): Likewise.

	* g++.dg/cpp26/decomp11.C: New test.